+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 - Fixes the link in the gist view will be opened in the iFrame.
+- Use the `postMessage` mechanism to set the iFrame's height, which should be much secure.
 
 ## 0.6.0